# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products/{id}": {
      "get": {
        "operationId": "ProductsController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Joriy sotuvchining bitta mahsulotini olish",
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Joriy sotuvchining mahsulotini tahrirlash",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "12",
                    "deleted": true
                  }
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Joriy sotuvchining mahsulotini soft-delete qilish",
        "tags": [
          "products"
        ]
      }
    },
...
    "/api/v1/storefront/products/{id}": {
      "get": {
        "operationId": "StorefrontController_getProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorefrontProductDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Faol mahsulot tafsiloti",
        "tags": [
          "storefront"
        ]
      }
    },
...
    "/api/v1/admin/products": {
      "post": {
        "operationId": "AdminProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Admin tanlangan do‘kon uchun yangi mahsulot yaratadi",
        "tags": [
          "admin-products"
        ]
      },
      "get": {
        "operationId": "AdminProductsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "ACTIVE",
                "ARCHIVED",
                "OUT_OF_STOCK"
              ]
            }
          },
          {
            "name": "blocked",
            "required": false,
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "shopId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "15",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "telefon",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "example": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Admin uchun barcha mahsulotlar va moderatsiya filtrlari",
        "tags": [
          "admin-products"
        ]
      }
    },
    "/api/v1/admin/products/{id}": {
      "get": {
        "operationId": "AdminProductsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-products"
        ]
      }
    },
    "/api/v1/admin/products/{id}/suspend": {
      "post": {
        "operationId": "AdminProductsController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-products"
        ]
      }
    },
    "/api/v1/admin/products/{id}/hide": {
      "post": {
        "operationId": "AdminProductsController_hide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HideProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mahsulotni sabab bilan storefront va searchdan yashirish",
        "tags": [
          "admin-products"
        ]
      }
    },
    "/api/v1/admin/products/{id}/reactivate": {
      "post": {
        "operationId": "AdminProductsController_reactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-products"
        ]
      }
    },
...
  },
  "info": {
    "title": "Elchi Marketplace API",
    "description": "Marketplace API — MVP (batafsil: docs/API_CONTRACT.md)",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorDetailDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": ""
          },
          "error": {
            "type": "string",
            "example": "phone +998XXXXXXXXX formatida bo‘lishi kerak"
          }
        },
        "required": [
          "field",
          "error"
        ]
      },
      "AuthErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "oneOf": [
              {
                "type": "string",
                "example": "Telefon yoki parol xato"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "password must be longer than or equal to 8 characters"
                ]
              }
            ]
          },
          "errorCode": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorDetailDto"
            }
          }
        },
        "required": [
          "statusCode",
          "message",
          "errorCode"
        ]
      },
...
      "SellerShopDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "15"
          },
          "ownerUserId": {
            "type": "string",
            "example": "42"
          },
          "name": {
            "type": "string",
            "example": "Ali Market"
          },
          "slug": {
            "type": "string",
            "example": "ali-market-1234567"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACTIVE",
              "SUSPENDED",
              "REJECTED"
            ],
            "example": "PENDING"
          },
          "description": {
            "type": "object",
            "example": "Maishiy texnika do‘koni",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "bannerUrl": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "phone": {
            "type": "object",
            "example": "+998901234567",
            "nullable": true
          },
          "regionId": {
            "type": "object",
            "example": "1",
            "nullable": true
          },
          "districtId": {
            "type": "object",
            "example": "10",
            "nullable": true
          },
          "address": {
            "type": "object",
            "example": "Toshkent shahri",
            "nullable": true
          },
          "rating": {
            "type": "number",
            "example": 0
          },
          "ordersCount": {
            "type": "number",
            "example": 0
          },
          "isFeatured": {
            "type": "boolean",
            "example": false
          },
          "tariffHome": {
            "type": "number",
            "example": 25000,
            "description": "Uy manziliga yetkazish tarifi"
          },
          "tariffCenter": {
            "type": "number",
            "example": 15000,
            "description": "Elchi markaziga yetkazish tarifi"
          }
        },
        "required": [
          "id",
          "ownerUserId",
          "name",
          "slug",
          "status",
          "description",
          "logoUrl",
          "bannerUrl",
          "phone",
          "regionId",
          "districtId",
          "address",
          "rating",
          "ordersCount",
          "isFeatured",
          "tariffHome",
          "tariffCenter"
        ]
      },
...
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "12"
          },
          "shopId": {
            "type": "string",
            "example": "5"
          },
          "ownerUserId": {
            "type": "string",
            "example": "42"
          },
          "categoryId": {
            "type": "object",
            "example": "1",
            "nullable": true
          },
          "name": {
            "type": "string",
            "example": "iPhone 16 Pro"
          },
          "slug": {
            "type": "string",
            "example": "iphone-16-pro"
          },
          "description": {
            "type": "object",
            "example": "Titanium, 256 GB",
            "nullable": true
          },
          "price": {
            "type": "number",
            "example": 14999000
          },
          "oldPrice": {
            "type": "object",
            "example": 15999000,
            "nullable": true
          },
          "imageUrl": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "attributes": {
            "type": "object",
            "example": {
              "brand": "Apple"
            }
          },
          "hasVariants": {
            "type": "boolean",
            "example": false
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "ARCHIVED",
              "OUT_OF_STOCK"
            ]
          },
          "isBlocked": {
            "type": "boolean",
            "example": false
          },
          "rating": {
            "type": "number",
            "example": 4.75
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-14T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-14T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "shopId",
          "ownerUserId",
          "categoryId",
          "name",
          "slug",
          "description",
          "price",
          "oldPrice",
          "imageUrl",
          "images",
          "attributes",
          "hasVariants",
          "status",
          "isBlocked",
          "rating",
          "createdAt",
          "updatedAt"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "iPhone 16 Pro"
          },
          "categoryId": {
            "type": "object",
            "example": "1",
            "nullable": true
          },
          "description": {
            "type": "object",
            "example": "Titanium, 256 GB"
          },
          "price": {
            "type": "number",
            "example": 14999000,
            "minimum": 0
          },
          "oldPrice": {
            "type": "object",
            "example": 15999000,
            "minimum": 0,
            "nullable": true
          },
          "imageUrl": {
            "type": "object",
            "example": "https://cdn.example.com/iphone.jpg",
            "description": "Asosiy rasm URL (yoki images massivida kamida bitta rasm bo‘lishi shart)"
          },
          "images": {
            "example": [
              "https://cdn.example.com/iphone-1.jpg"
            ],
            "description": "Mahsulot rasmlari (agar imageUrl bo‘sh bo‘lsa, kamida bitta rasm talab qilinadi)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "attributes": {
            "type": "object",
            "example": {
              "brand": "Apple",
              "storage": "256 GB"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "ARCHIVED",
              "OUT_OF_STOCK"
            ],
            "default": "ACTIVE"
          }
        }
      },
      "StorefrontCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "7"
          },
          "name": {
            "type": "string",
            "example": "Smartfonlar"
          },
          "slug": {
            "type": "string",
            "example": "smartfonlar"
          },
          "parentId": {
            "type": "object",
            "example": "2",
            "nullable": true
          },
          "iconUrl": {
            "type": "object",
            "example": null,
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "parentId",
          "iconUrl"
        ]
      },
      "ProductVariantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "25"
          },
          "productId": {
            "type": "string",
            "example": "35"
          },
          "sku": {
            "type": "string",
            "example": "IPHONE-16-BLACK-256"
          },
          "name": {
            "type": "object",
            "example": "Qora-256 Gb",
            "nullable": true
          },
          "attributes": {
            "type": "object",
            "example": {
              "color": "Qora",
              "storage": "256 Gb"
            }
          },
          "price": {
            "type": "object",
            "example": 16000000,
            "nullable": true
          },
          "oldPrice": {
            "type": "object",
            "example": 17000000,
            "nullable": true
          },
          "barcode": {
            "type": "object",
            "example": "4780012345678",
            "nullable": true
          },
          "imageUrl": {
            "type": "object",
            "example": "https://cdn.example.com/products/iphone-black.jpg",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "productId",
          "sku",
          "name",
          "attributes",
          "price",
          "oldPrice",
          "barcode",
          "imageUrl",
          "isActive"
        ]
      },
      "StorefrontProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "12"
          },
          "shopId": {
            "type": "string",
            "example": "5"
          },
          "ownerUserId": {
            "type": "string",
            "example": "42"
          },
          "categoryId": {
            "type": "object",
            "example": "1",
            "nullable": true
          },
          "name": {
            "type": "string",
            "example": "iPhone 16 Pro"
          },
          "slug": {
            "type": "string",
            "example": "iphone-16-pro"
          },
          "description": {
            "type": "object",
            "example": "Titanium, 256 GB",
            "nullable": true
          },
          "price": {
            "type": "number",
            "example": 14999000
          },
          "oldPrice": {
            "type": "object",
            "example": 15999000,
            "nullable": true
          },
          "imageUrl": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "attributes": {
            "type": "object",
            "example": {
              "brand": "Apple"
            }
          },
          "hasVariants": {
            "type": "boolean",
            "example": false
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "ARCHIVED",
              "OUT_OF_STOCK"
            ]
          },
          "isBlocked": {
            "type": "boolean",
            "example": false
          },
          "rating": {
            "type": "number",
            "example": 4.75
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-14T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-14T10:00:00.000Z"
          },
          "shop": {
            "description": "Product qaysi marketga tegishli ekanini bildiradi",
            "allOf": [
              {
                "$ref": "#/components/schemas/SellerShopDto"
              }
            ]
          },
          "category": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/StorefrontCategoryDto"
              }
            ]
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantDto"
            }
          }
        },
        "required": [
          "id",
          "shopId",
          "ownerUserId",
          "categoryId",
          "name",
          "slug",
          "description",
          "price",
          "oldPrice",
          "imageUrl",
          "images",
          "attributes",
          "hasVariants",
          "status",
          "isBlocked",
          "rating",
          "createdAt",
          "updatedAt",
          "shop",
          "category",
          "variants"
        ]
      },
...
      "AdminCreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "iPhone 16 Pro"
          },
          "categoryId": {
            "type": "object",
            "example": "1",
            "nullable": true
          },
          "description": {
            "type": "object",
            "example": "Titanium, 256 GB"
          },
          "price": {
            "type": "number",
            "example": 14999000,
            "minimum": 0
          },
          "oldPrice": {
            "type": "object",
            "example": 15999000,
            "minimum": 0,
            "nullable": true
          },
          "imageUrl": {
            "type": "object",
            "example": "https://cdn.example.com/iphone.jpg",
            "description": "Asosiy rasm URL (yoki images massivida kamida bitta rasm bo‘lishi shart)"
          },
          "images": {
            "example": [
              "https://cdn.example.com/iphone-1.jpg"
            ],
            "description": "Mahsulot rasmlari (agar imageUrl bo‘sh bo‘lsa, kamida bitta rasm talab qilinadi)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "attributes": {
            "type": "object",
            "example": {
              "brand": "Apple",
              "storage": "256 GB"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "ARCHIVED",
              "OUT_OF_STOCK"
            ],
            "default": "ACTIVE"
          },
          "shopId": {
            "type": "string",
            "example": "15",
            "description": "Mahsulot biriktiriladigan do‘kon IDsi"
          }
        },
        "required": [
          "name",
          "price",
          "shopId"
        ]
      },
      "HideProductDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Mahsulot rasmi yoki tavsifi qoidalarga zid"
          }
        },
        "required": [
          "reason"
        ]
      },
...
    }
  }
}